+++ docs/main.js
@@ -1,6 +1,7 @@
 /**
  * @typedef {Object} Configuration
  * @memberof module:main
+ * @property {module:tasks/babel.Parameters} [babel] Parameters of the babel task
  * @property {module:tasks/checkFileNames.Parameters} [checkFileNames] Parameters of the checkFileNames task
  * @property {module:tasks/jsdoc.Parameters} [jsdoc] Parameters of the jsdoc task
  * @property {module:tasks/nsp.Parameters} [nsp] Parameters of the nsp task
@@ -16,9 +17,9 @@
  * @typedef {Object} GlobalConfiguration
  * @memberof module:main
  * @property {Object} babel Common babel configuration, used by:
- *                          jest, jsdoc, webpack
+ *                          babel, jest, jsdoc, webpack
  * @property {string|string[]} sourceFiles Common source files, used by:
- *                                         eslint, jscs, check-dependencies
+ *                                         babel, eslint, jscs, check-dependencies
  * @property {boolean} allowLinking Whether to allow linking globally installed packages, used by:
  *                                  npm-install
  */